Remove SUDS (session-user-data-service)
Coral is now our source of truth as we are decommissioning SUDS. This PR
removes SUDS and as a result, also removes the `/hotarticles` route.
Somehow this is connected with alphaville-blogs and renders this page:
https://ftalphaville.ft.com/most-popular-livefyre.

I can't find where this URL is being linked to so am unsure how a user
would even find this page :shrug:

There is are two links n the Alphaville drawer: Most Commented and Most
Popular however those aren't related and were purposely deleted. See
https://github.com/Financial-Times/alphaville-blogs/commit/330e434c581f67772b2a5223ebd201c63596af17.